# Optimize TryReadUntil