--- a/src/System.Buffers.ReaderWriter/System/Buffers/Reader/BufferReader_search.cs
+++ b/src/System.Buffers.ReaderWriter/System/Buffers/Reader/BufferReader_search.cs
@@ -3,128 +3,179 @@
 
 namespace System.Buffers.Reader
 {
-    // TODO: the TryReadUntil methods are very inneficient. We need to fix that.
-    public static partial class BufferReaderExtensions
+    public ref partial struct BufferReader
     {
-        public static bool TryReadUntil(ref BufferReader reader, out ReadOnlySpan<byte> bytes, byte delimiter)
+        /// <summary>
+        /// Try to read everything up to the given delimiter. Will position the reader past the delimiter if found.
+        /// </summary>
+        /// <param name="span">The read data, if any.</param>
+        /// <param name="delimiter">The delimiter to look for.</param>
+        /// <returns>True if the data was found.</returns>
+        public bool TryReadUntil(out ReadOnlySpan<byte> span, byte delimiter)
         {
-            int idx = reader.UnreadSegment.IndexOf(delimiter);
-            if (idx != -1)
+            ReadOnlySpan<byte> remaining = CurrentSegmentIndex == 0 ? CurrentSegment : UnreadSegment;
+            int index = remaining.IndexOf(delimiter);
+            if (index != -1)
             {
-                bytes = reader.UnreadSegment.Slice(0, idx);
-                reader.Advance(idx + 1);
+                span = index == 0 ? default : remaining.Slice(0, index);
+                Advance(index + 1);
                 return true;
             }
-            else
-            {
-                bool result = TryReadUntil(ref reader, out ReadOnlySequence<byte> sequence, delimiter);
 
-                if (sequence.IsSingleSegment)
-                {
-                    bytes = sequence.First.Span;
-                }
-                else
-                {
-                    bytes = sequence.ToArray();
-                }
-
-                return result;
-            }
+            return TryReadUntilSlow(out span, delimiter, remaining.Length);
         }
 
-        public static bool TryReadUntil(ref BufferReader reader, out ReadOnlySequence<byte> bytes, byte delimiter)
+        private bool TryReadUntilSlow(out ReadOnlySpan<byte> span, byte delimiter, int advance)
         {
-            BufferReader copy = reader;
-            while (!reader.End)
+            Advance(advance);
+            if (End || !TryReadUntil(out ReadOnlySequence<byte> sequence, delimiter))
             {
-                if (reader.Peek() == delimiter)
+                span = default;
+                return false;
+            }
+
+            span = sequence.IsSingleSegment ? sequence.First.Span : sequence.ToArray();
+            return true;
+        }
+
+        /// <summary>
+        /// Try to read everything up to the given delimiter. Will position the reader past the delimiter if found.
+        /// </summary>
+        /// <param name="sequence">The read data, if any.</param>
+        /// <param name="delimiter">The delimiter to look for.</param>
+        /// <returns>True if the data was found.</returns>
+        public bool TryReadUntil(out ReadOnlySequence<byte> sequence, byte delimiter)
+        {
+            BufferReader copy = this;
+            ReadOnlySpan<byte> remaining = CurrentSegmentIndex == 0 ? CurrentSegment : UnreadSegment;
+
+            while (!End)
+            {
+                int index = remaining.IndexOf(delimiter);
+                if (index != -1)
                 {
-                    bytes = reader.Sequence.Slice(copy.Position, reader.Position);
-                    reader.Advance(1);
+                    // Found the delimiter. Move to it, slice, then move past it.
+                    if (index > 0)
+                    {
+                        Advance(index);
+                    }
+
+                    sequence = Sequence.Slice(copy.Position, Position);
+                    Advance(1);
                     return true;
                 }
-                reader.Advance(1);
+
+                Advance(remaining.Length);
+                remaining = CurrentSegment;
             }
-            reader = copy;
-            bytes = default;
+
+            // Didn't find anything, reset our original state.
+            this = copy;
+            sequence = default;
             return false;
         }
 
-        public static bool TryReadUntilAny(ref BufferReader reader, out ReadOnlySpan<byte> bytes, ReadOnlySpan<byte> delimiters)
+        /// <summary>
+        /// Try to read everything up to the given delimiters. Will position the reader past the delimiter if found.
+        /// </summary>
+        /// <param name="span">The read data, if any.</param>
+        /// <param name="delimiters">The delimiters to look for.</param>
+        /// <returns>True if the data was found.</returns>
+        public bool TryReadUntilAny(out ReadOnlySpan<byte> span, ReadOnlySpan<byte> delimiters)
         {
-            int idx = reader.UnreadSegment.IndexOfAny(delimiters);
-            if (idx != -1)
+            ReadOnlySpan<byte> remaining = UnreadSegment;
+            int index = remaining.IndexOfAny(delimiters);
+            if (index != -1)
             {
-                bytes = reader.UnreadSegment.Slice(0, idx);
-                reader.Advance(idx);
+                span = remaining.Slice(0, index);
+                Advance(index + 1);
                 return true;
             }
-            else
-            {
-                bool result = TryReadUntil(ref reader, out ReadOnlySequence<byte> sequence, delimiters);
 
-                if (sequence.IsSingleSegment)
-                {
-                    bytes = sequence.First.Span;
-                }
-                else
-                {
-                    bytes = sequence.ToArray();
-                }
-
-                return result;
-            }
+            return TryReadUntilAnySlow(out span, delimiters, remaining.Length);
         }
 
-        private static bool TryReadUntilAny(ref BufferReader reader, out ReadOnlySequence<byte> bytes, ReadOnlySpan<byte> delimiters)
+        private bool TryReadUntilAnySlow(out ReadOnlySpan<byte> span, ReadOnlySpan<byte> delimiters, int advance)
         {
-            BufferReader copy = reader;
-            while (!reader.End)
+            Advance(advance);
+            if (End || !TryReadUntilAny(out ReadOnlySequence<byte> sequence, delimiters))
             {
-                byte b = (byte)reader.Peek();
-                if (delimiters.IndexOf(b) != -1)
+                span = default;
+                return false;
+            }
+
+            span = sequence.IsSingleSegment ? sequence.First.Span : sequence.ToArray();
+            return true;
+        }
+
+        /// <summary>
+        /// Try to read everything up to the given delimiters. Will position the reader past the delimiter if found.
+        /// </summary>
+        /// <param name="sequence">The read data, if any.</param>
+        /// <param name="delimiters">The delimiters to look for.</param>
+        /// <returns>True if the data was found.</returns>
+        public bool TryReadUntilAny(out ReadOnlySequence<byte> sequence, ReadOnlySpan<byte> delimiters)
+        {
+            BufferReader copy = this;
+            ReadOnlySpan<byte> remaining = CurrentSegmentIndex == 0 ? CurrentSegment : UnreadSegment;
+
+            while (!End)
+            {
+                int index = remaining.IndexOfAny(delimiters);
+                if (index != -1)
                 {
-                    bytes = reader.Sequence.Slice(copy.Position, reader.Position);
+                    // Found one of the delimiters. Move to it, slice, then move past it.
+                    if (index > 0)
+                    {
+                        Advance(index);
+                    }
+
+                    sequence = Sequence.Slice(copy.Position, Position);
+                    Advance(1);
                     return true;
                 }
-                reader.Advance(1);
+
+                Advance(remaining.Length);
+                remaining = CurrentSegment;
             }
-            reader = copy;
-            bytes = default;
+
+            // Didn't find anything, reset our original state.
+            this = copy;
+            sequence = default;
             return false;
         }
 
-        public static bool TryReadUntil(ref BufferReader reader, out ReadOnlySequence<byte> bytes, ReadOnlySpan<byte> delimiter)
+        public bool TryReadUntil(out ReadOnlySequence<byte> sequence, ReadOnlySpan<byte> delimiter)
         {
             if (delimiter.Length == 0)
             {
-                bytes = default;
+                sequence = default;
                 return true;
             }
 
             int matched = 0;
-            var copy = reader;
-            var start = reader.Position;
-            var end = reader.Position;
-            while (!reader.End)
+            BufferReader copy = this;
+            SequencePosition start = Position;
+            SequencePosition end = Position;
+            while (!End)
             {
-                if (reader.Read() == delimiter[matched])
+                if (Read() == delimiter[matched])
                 {
                     matched++;
                 }
                 else
                 {
-                    end = reader.Position;
+                    end = Position;
                     matched = 0;
                 }
                 if (matched >= delimiter.Length)
                 {
-                    bytes = reader.Sequence.Slice(start, end);
+                    sequence = Sequence.Slice(start, end);
                     return true;
                 }
             }
-            reader = copy;
-            bytes = default;
+            this = copy;
+            sequence = default;
             return false;
         }
     }