Move the methods to instance and optimize. I did not optimize the TryReadUntil that takes a Span delimiter yet.

For a single segment single delimiter I can get about 25% gain. The cost of comparing multiple delimiters eats a greater portion of the time pretty rapidly. I think were pretty close to optimal here (i.e. most of the time is in IndexOf/Any).

**Before (single segment)**
```
                            Method |       Mean |     Error |    StdDev |
---------------------------------- |-----------:|----------:|----------:|
      TryReadUntil_Span_OneSegment |   179.8 us |  3.220 us |  2.854 us |
 TryReadUntilAny_2_Span_OneSegment |   633.3 us |  6.364 us |  5.953 us |
 TryReadUntilAny_5_Span_OneSegment | 3,116.2 us | 16.157 us | 15.113 us |
```

**After (single segment)**
```
                            Method |       Mean |     Error |    StdDev |
---------------------------------- |-----------:|----------:|----------:|
      TryReadUntil_Span_OneSegment |   142.5 us |  2.651 us |  2.479 us |
 TryReadUntilAny_2_Span_OneSegment |   616.5 us |  6.070 us |  5.678 us |
 TryReadUntilAny_5_Span_OneSegment | 3,104.0 us | 51.189 us | 47.882 us |
```

For multiple segments I can get better improvements.

**Before (multiple segments)**
```
                     Method |     Mean |     Error |    StdDev |
--------------------------- |---------:|----------:|----------:|
      TryReadUntil_Sequence | 2.559 ms | 0.0390 ms | 0.0365 ms |
 TryReadUntilAny_2_Sequence | 5.848 ms | 0.0439 ms | 0.0390 ms |
 TryReadUntilAny_5_Sequence | 6.537 ms | 0.0312 ms | 0.0292 ms |
          TryReadUntil_Span | 3.278 ms | 0.0149 ms | 0.0133 ms |
     TryReadUntilAny_2_Span | 6.348 ms | 0.0305 ms | 0.0285 ms |
     TryReadUntilAny_5_Span | 8.161 ms | 0.0414 ms | 0.0367 ms |
```

**After (multiple segments)**
```
                     Method |     Mean |     Error |    StdDev |
--------------------------- |---------:|----------:|----------:|
      TryReadUntil_Sequence | 1.278 ms | 0.0051 ms | 0.0045 ms |
 TryReadUntilAny_2_Sequence | 2.185 ms | 0.0115 ms | 0.0089 ms |
 TryReadUntilAny_5_Sequence | 5.009 ms | 0.0486 ms | 0.0455 ms |
          TryReadUntil_Span | 1.817 ms | 0.0139 ms | 0.0130 ms |
     TryReadUntilAny_2_Span | 2.906 ms | 0.0104 ms | 0.0092 ms |
     TryReadUntilAny_5_Span | 5.548 ms | 0.0402 ms | 0.0376 ms |
```

cc: @benaadams, @mgravell, @Drawaes