@JeremyKuhne 2018-08-08
I think I inadvertently broke the JSON reader by making the methods consistent. They weren't all moving past the delimiter. Investigating...